Refactor components/functions to separate files
